pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asgard_camera_py"
version = "0.1.0"
description = "Python wrapper for the Heimdallr/Baldr C-RED One camera with Asgard system integration and shared memory access."
readme = "README.md"
requires-python = ">=3.8"

authors = [
    { name = "benjamin courtney-barrer", email = "bencb92@gmail.com" }
]

license = { text = "MIT" }

dependencies = [
    "numpy>=1.20",
    "astropy>=5.0",
    "pyzmq>=24.0",
    "xaosim>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/courtney-barrer/asgard_camera_py"
Repository = "https://github.com/courtney-barrer/asgard_camera_py"
